Some more phpstan errors (task #8244)

+++ src/Model/Table/ArticlesTable.php
@@ -15,6 +15,7 @@ use ArrayObject;
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use Cake\Datasource\EntityInterface;
+use Cake\Datasource\ResultSetInterface;
 use Cake\Event\Event;
 use Cake\I18n\Time;
 use Cake\ORM\Entity;
@@ -31,6 +32,7 @@ use InvalidArgumentException;
 /**
  * Articles Model
  *
+ * @property \Cake\ORM\Association\HasMany $ArticleFeaturedImages
  */
 class ArticlesTable extends Table
 {
@@ -62,7 +64,7 @@ class ArticlesTable extends Table
     /**
      * Article searchable fields.
      *
-     * @var string
+     * @var mixed[]
      */
     protected $_searchableFields = ['title', 'excerpt', 'content'];
 
@@ -83,7 +85,7 @@ class ArticlesTable extends Table
         $this->addBehavior('Timestamp');
         $this->addBehavior('Muffin/Trash.Trash');
         $this->addBehavior('Muffin/Slug.Slug', [
-            'unique' => function (Entity $entity, $slug, $separator) {
+            'unique' => function (EntityInterface $entity, $slug, $separator) {
                 return $this->_uniqueSlug($entity, $slug, $separator);
             }
         ]);
@@ -214,7 +216,12 @@ class ArticlesTable extends Table
             }
 
             foreach ($shortcodes as $shortcode) {
-                $cacheKey = 'shortcode_' . md5(json_encode($shortcode));
+                $encoded = json_encode($shortcode);
+                if (!$encoded) {
+                    continue;
+                }
+
+                $cacheKey = 'shortcode_' . md5($encoded);
                 // delete shortcode cache
                 Cache::delete($cacheKey);
             }
@@ -226,7 +233,7 @@ class ArticlesTable extends Table
      *
      * @return string
      */
-    public function getSearchQuery()
+    public function getSearchQuery(): string
     {
         return $this->_searchQuery;
     }
@@ -235,9 +242,10 @@ class ArticlesTable extends Table
      * Search query setter.
      *
      * @param string $searchQuery Search query string
+     *
      * @return void
      */
-    public function setSearchQuery(string $searchQuery)
+    public function setSearchQuery(string $searchQuery): void
     {
         if (!is_string($searchQuery)) {
             throw new InvalidArgumentException('Search query must be a string.');
@@ -251,9 +259,10 @@ class ArticlesTable extends Table
      *
      * @param \Cake\ORM\Query $query Query instance
      * @param string $searchQuery Search query value
+     *
      * @return void
      */
-    public function applySearch(Query $query, $searchQuery)
+    public function applySearch(Query $query, string $searchQuery): void
     {
         if (empty($searchQuery)) {
             return;
@@ -271,13 +280,15 @@ class ArticlesTable extends Table
      * Fetch and return Article by id or slug.
      *
      * @param string $id Article id or slug.
-     * @param string $siteId Site id.
+     * @param string|null $siteId Site id.
      * @param bool $associated Contain associated articles and images.
-     * @return \Cake\ORM\Entity
+     *
      * @throws \Cake\Datasource\Exception\RecordNotFoundException
      * @throws \InvalidArgumentException
+     *
+     * @return \Cake\Datasource\EntityInterface
      */
-    public function getArticle(string $id, $siteId = null, $associated = false)
+    public function getArticle(string $id, ?string $siteId, bool $associated = false): EntityInterface
     {
         if (empty($id)) {
             throw new InvalidArgumentException('Article id or slug cannot be empty.');
@@ -301,12 +312,18 @@ class ArticlesTable extends Table
                     'Articles.id' => $id,
                     'Articles.slug' => $id
                 ]
-            ])
-            ->contain($contain)
-            ->limit(1)
-            ->applyOptions(['site_id' => $siteId]);
+            ]);
+        $query->enableHydration(true);
+        $query->limit(1);
+        $query->contain($contain);
+        $query->applyOptions(['site_id' => $siteId]);
 
-        return $query->firstOrFail();
+        /**
+         * @var \Cake\Datasource\EntityInterface
+         */
+        $result = $query->firstOrFail();
+
+        return $result;
     }
 
     /**
@@ -321,7 +338,7 @@ class ArticlesTable extends Table
      *
      * @return \Cake\Datasource\ResultSetInterface
      */
-    public function getArticles(string $siteId, string $type, bool $associated = false)
+    public function getArticles(string $siteId, string $type, bool $associated = false): ResultSetInterface
     {
         if (!is_string($siteId)) {
             throw new InvalidArgumentException('Site id or slug must be a string.');
@@ -348,20 +365,27 @@ class ArticlesTable extends Table
         }
 
         $query = $this->find('all')
-            ->where($conditions)
-            ->contain($contain)
-            ->applyOptions(['site_id' => $siteId]);
+            ->where($conditions);
+        $query->enableHydration(true);
+        $query->contain($contain);
+        $query->applyOptions(['site_id' => $siteId]);
 
-        return $query->all();
+        /**
+         * @var \Cake\Datasource\ResultSetInterface
+         */
+        $result = $query->all();
+
+        return $result;
     }
 
     /**
      * Returns supported types.
      *
      * @param bool $withOptions Flag for including type options
-     * @return array
+     *
+     * @return mixed[]
      */
-    public function getTypes($withOptions = true)
+    public function getTypes(bool $withOptions = true): array
     {
         if (!empty($this->_types)) {
             return $this->_types;
@@ -393,7 +417,7 @@ class ArticlesTable extends Table
             $v['label'] = $v['label'] ? $v['label'] : Inflector::humanize($k);
         }
 
-        if (!(bool)$withOptions) {
+        if (! $withOptions) {
             return array_keys($this->_types);
         }
 
@@ -404,7 +428,7 @@ class ArticlesTable extends Table
      * Returns type options.
      *
      * @param string $type Type name
-     * @return array
+     * @return mixed[]
      */
     public function getTypeOptions(string $type): array
     {
@@ -423,27 +447,34 @@ class ArticlesTable extends Table
     /**
      * Returns specified category(ies) articles.
      *
-     * @param array $ids Category(ies) ID(s)
+     * @param mixed[] $ids Category(ies) ID(s)
      * @return \Cake\Datasource\ResultSetInterface
      */
-    public function getArticlesByCategory(array $ids)
+    public function getArticlesByCategory(array $ids): ResultSetInterface
     {
         $query = $this->find('all')
-            ->where(['Articles.category_id IN' => $ids])
-            ->contain(['ArticleFeaturedImages']);
+            ->where(['Articles.category_id IN' => $ids]);
+        $query->enableHydration(true);
+        $query->contain(['ArticleFeaturedImages']);
 
-        return $query->all();
+        /**
+         * @var \Cake\Datasource\ResultSetInterface
+         */
+        $result = $query->all();
+
+        return $result;
     }
 
     /**
      * Returns a unique slug.
      *
-     * @param \Cake\ORM\Entity $entity Entity.
+     * @param \Cake\Datasource\EntityInterface $entity Entity.
      * @param string $slug Slug.
      * @param string $separator Separator.
+     *
      * @return string Unique slug.
      */
-    protected function _uniqueSlug(Entity $entity, $slug, $separator)
+    protected function _uniqueSlug(EntityInterface $entity, string $slug, string $separator): string
     {
         $behavior = $this->behaviors()->Slug;
